Add autopopulate, date/time autofill
When advancing from the event detail page, perform an AJAX request to pre-fill the services page with the details of a previous event, if the server finds a suitable previous event.
When the user enters a date or time in the event details page, automatically fill any empty dates/times below that one.
Fix body being unnecessarily tall
Add prices to service cards
Do not list N/A items in review page